Backgrounds

The background() function adjusts the background color of a tag element.

background(x, color)

Arguments

x
A tag element or .style pronoun.

color
One of "blue", "indigo", "purple", "red", "orange", "yellow", "green", "teal", "cyan", "white", or "transparent" specifying the background color of the tag element.

Buttons

Use background() to modify shiny’s action buttons.

actionButton("id", "Take action") %>%
  background("green")

With a couple other functions we can take our improvement a step further.

actionButton("id", "Take action") %>%
  background("green") %>%
  border("green") %>%
  shadow("small")

Shiny’s download buttons include ..., so we can use the .style pronoun!

downloadButton(
  .style %>%
    background("white") %>%
    border("blue"),
  outputId = "dwnld",
  label = "Do a download",
  class = NULL
)

Panels

sidebarLayout(
  sidebarPanel(
    .style %>%
      background("blue"),
    "It's alive"
  ),
  mainPanel(
    .style %>%
      background("red"),
    "It's panel"
  )
)

It’s panel

Colors

blue

indigo

purple

red

orange

yellow

green

teal

cyan

white

transparent

Examples

library(htmltools)

div(
  .style %>%
    background("white") %>%
    border("blue") %>%
    text("white"),
  "Nunc porta vulputate tellus.",
  "Suspendisse potenti."
)

#> <div class="cas-background-white cas-border-blue cas-border-all cas-border-round-md cas-text-color-white">
#>   Nunc porta vulputate tellus.
#>   Suspendisse potenti.
#> </div>