Borders

The border() function adjusts a tag element's borders.

border(x, color, sides = TRUE, width = 1, round = "medium")

Arguments

x

A tag element or .style pronoun.

color

One of "blue", "indigo", "purple", "red", "orange", "yellow", "green", "teal", "cyan", or "white" specifying the border color.

sides

One or more of "top", "t", "right", "r", "bottom", "b", "left", "l", "all", or "none" specifying which sides to add borders to, defaults to TRUE. TRUE and FALSE may be used as shorthand for all sides or no sides, respectively.

width

One or more of 1, 2, 3, 4, or 5 specifying the width of the element's border, defaults to 1.

round

One of "small", "sm", "medium", "md", "large", "lg", or "none" specifying how to round the corners of the element, defaults to "medium".

Colors

blue

indigo

purple

red

orange

yellow

green

teal

cyan

white

Round

small

sm

medium

md

large

lg

none

Examples

library(htmltools)

h3(
  .style %>%
    border("red", "bottom") %>%
    text("red"),
  "A bright, underlined heading!"
)
#> <h3 class="cas-border-bottom cas-border-red cas-border-round-md cas-text-red">A bright, underlined heading!</h3>