Display

The display() function adjusts how a tag element is rendered. For example, to use the flex box layout the display must be "flex".

display(x, type)

Arguments

x
A tag element or .style pronoun.

type
A responsive argument.

One of "inline", "inline-block", "block", "table", "table-row", "table-cell", "flex", "inline-flex", or "none".

Block vs inline

div(
  div(
    .style %&gt;%
      border("blue"),
    "block"
  ),
  div(
    .style %&gt;%
      border("blue"),
    "block"
  )
)

block

block

div(
  div(
    .style %&gt;%
      border("blue") %&gt;%
      display("inline"),
    "inline"
  ),
  div(
    .style %&gt;%
      border("blue") %&gt;%
      display("inline"),
    "inline"
  )
)

inline

inline

Examples

library(htmltools)

# When using flex make sure you specify the flex display.
div(
  .style %>%
    display("flex") %>%
    flex(justify = "center"),
  "Powerful stuff"
)

#> <div class="cas-display-flex cas-flex-row cas-flex-content-center cas-flex-items-stretch cas-flex-nowrap">Powerful stuff</div>