Fixed position elements

The fixed() function affixes an element to the top or bottom of the page. Because of their nature fixed elements may cover up other elements on the page. Use padding() or margin() to adjust any covered elements.

fixed(x, position)

Arguments

x
A tag element or .style pronoun.

position
One of "top" or "bottom" specifying the where to fix the element on the page.

Examples

library(htmltools)

tags$nav(
  .style %>%
    fixed("top") %>%
    background("blue"),
  tags$a(href = "home.html", "Home")
)

#> <nav class="cas-fixed-top cas-background-blue">
#>   <a href="home.html">Home</a>
#> </nav>