Flex

The flex() function adjusts the flex box layout of an element. To use the flex box layout the element must also use the flex display, see display(). The flex box layout is incredibly powerful and allows centering of elements vertically or horizontally, automatic adjustment of space between or around child elements, and more.

Direct child elements of a flex box container are automatically considered flex items and may be adjusted with item().

flex(x, direction = "row", justify = "start", align = "stretch", wrap = FALSE)

Arguments

x
A tag element or .style pronoun.

direction
A responsive argument.

One of "row" or "column" specifying the main axis of flex items, defaults to "row".

If "row", the main axis is horizontal and items are arranged from left to right. The cross axis is the vertical.

If "column", the main axis is vertical and items are arranged from top to bottom. The cross axis is the horizontal.

justify
A responsive argument.

One of "start", "end", "center", "around", or "between" specifying how items are arranged on the main axis, defaults to "start".

If "between" or "around", items are arranged by evenly sharing the space between or around the items.

align
A responsive argument.

One of "start", "end", "center", "baseline", or "stretch" specifying how items are arranged on the cross axis, defaults to "stretch".

wrap
A responsive argument.

One of TRUE or FALSE specifying if items are forced onto one line or allowed to wrap onto multiple lines, defaults to FALSE.

Centering elements

Center an input above a larger element or next to the element if space allows.

div(
  .style %&gt;%
    display("flex") %&gt;%
    flex(direction = c(default = "column", md = "row")),

  radioButtons("id", "A sample input", c("Choice 1", "Choice 2")) %&gt;%
    margin(h = c(xs = "auto", md = 0)),

  div(
    .style %&gt;%
      width(100) %&gt;%
      background("indigo") %&gt;%
      text("white"),
    "Plot placeholder"
  )
)

A sample input

Choice 1

Choice 2

Plot placeholder

Details

This section needs pretty specific examples of how to use flex. I don’t know that people will want a tutorial on flex.

For the sake of the demo let’s create a flex item help function.

flexItem &lt;- function(...) {
  div(
    .style %&gt;% padding(3) %&gt;% border("blue"),
    "A flex item",
    ...
  )
}

Different directions

Many of flex()’s arguments are viewport responsive. On small screens the flex items are placed vertically and can occupy the full width of the mobile device. On medium or larger screens the items are placed horizontally.

div(
  .style %&gt;%
    display("flex") %&gt;%
    flex(
      direction = c(xs = "column", md = "row")  # &lt;-
    ),
  flexItem(),
  flexItem(),
  flexItem()
)

A flex item

A flex item

A flex item

Resize the browser for this example.

You can keep items as a column by specifying only "column".

div(
  .style %&gt;%
    display("flex") %&gt;%
    flex(direction = "column"),  # &lt;-
  flexItem(),
  flexItem(),
  flexItem()
)

A flex item

A flex item

A flex item

Spacing items with justify

Below is a series of examples showing how to change the horizontal alignment of your flex items. Let’s start by pushing items to the beginning of their parent container.

div(
  .style %&gt;%
    display("flex") %&gt;%
    flex(justify = "start"),  # &lt;-
  flexItem(),
  flexItem(),
  flexItem(),
  flexItem()
)

A flex item

A flex item

A flex item

A flex item

We can also push items to the end.

div(
  .style %&gt;%
    display("flex") %&gt;%
    flex(justify = "end"),  # &lt;-
  flexItem(),
  flexItem(),
  flexItem(),
  flexItem()
)

A flex item

A flex item

A flex item

A flex item

Without using a table layout we can center items.

div(
  .style %&gt;%
    display("flex") %&gt;%
    flex(justify = "center"),  # &lt;-
  flexItem(),
  flexItem(),
  flexItem(),
  flexItem()
)

A flex item

A flex item

A flex item

A flex item

You can also put space between items

div(
  .style %&gt;%
    display("flex") %&gt;%
    flex(justify = "between"),  # &lt;-
  flexItem(),
  flexItem(),
  flexItem(),
  flexItem()
)

A flex item

A flex item

A flex item

A flex item

… or put space around items.

div(
  .style %&gt;%
    display("flex") %&gt;%
    flex(justify = "around"),  # &lt;-
  flexItem(),
  flexItem(),
  flexItem(),
  flexItem()
)

A flex item

A flex item

A flex item

A flex item

The "between" and "around" values come from the original CSS values "space-between" and "space-around".

wrap onto new lines

Using flexbox we can also control how items wrap onto new lines.

div(
  .style %&gt;%
    display("flex") %&gt;%
    flex(wrap = TRUE),
  flexItem(),
  flexItem(),
  flexItem(),
  flexItem()
)

A flex item

A flex item

A flex item

A flex item

Examples

library(htmltools)

div(
  .style %>%
    display("flex") %>%
    flex(justify = "end"),
  div("Aliquam posuere."),
  div("Lorem ipsum dolor sit amet, consectetuer adipiscing elit.")
)

#> <div class="cas-display-flex cas-flex-row cas-flex-content-end cas-flex-items-stretch cas-flex-nowrap">
#>   <div>Aliquam posuere.</div>
#>   <div>Lorem ipsum dolor sit amet, consectetuer adipiscing elit.</div>
#> </div>