Flex

The flex_*() functions adjust the flexbox layout of an element. The flexbox layout is incredibly powerful and allows centering of elements vertically and horizontally, automatic adjustment of space between and around child elements, and more. To use flexbox make sure to include flex_display() when styling an element. For finer control over an element's display see display().

Direct child elements of a flex box container are automatically considered flex items and may be adjusted with the item_*() functions, see item_align().

Usage

flex_display(x)

flex_direction(x, ...)

flex_justify(x, ...)

flex_align(x, ...)

flex_gap(x, ...)

flex_wrap(x, ...)

flex_content(x, ...)

Arguments

x

A tag element or .style pronoun.

...

Name-value pairs.

Examples

library(htmltools)

div(
  .style %>%
    flex_display() %>%
    flex_justify("end"),
  div("Item 1"),
  div("Item 2"),
  div("Item 3")
)
#> <div class="d-flex justify-content-end">
#>   <div>Item 1</div>
#>   <div>Item 2</div>
#>   <div>Item 3</div>
#> </div>