Flexbox

The flexbox() function adjusts the flex box layout of an element. To use the flex box layout the element must also use the flex display, see display(). The flex box layout is incredibly powerful and allows centering of elements vertically or horizontally, automatic adjustment of space between or around child elements, and more.

Direct child elements of a flex box container are automatically considered flex items and may be adjusted with flex().

flexbox(
  x,
  direction = "row",
  justify = "start",
  align = "stretch",
  wrap = FALSE
)

Arguments

x
A tag element or .style pronoun.

direction
A responsive argument.

One of "row" or "column" specifying the main axis of flex items, defaults to "row".

If "row", the main axis is horizontal and items are arranged from left to right. The cross axis is the vertical.

If "column", the main axis is vertical and items are arranged from top to bottom. The cross axis is the horizontal.

justify
A responsive argument.

One of "start", "end", "center", "around", or "between" specifying how items are arranged on the main axis, defaults to "start".

If "between" or "around", items are arranged by evenly sharing the space between or around the items.

align
A responsive argument.

One of "start", "end", "center", "baseline", or "stretch" specifying how items are arranged on the cross axis, defaults to "stretch".

wrap
A responsive argument.

One of TRUE or FALSE specifying if items are forced onto one line or allowed to wrap onto multiple lines, defaults to FALSE.