Floats

The float() function places an element to the left or right side of its parent element. Other text and inline elements wrap around floated elements.

float(x, side)

Arguments

x
A tag element or .style pronoun.

side
A responsive argument.

One of "left" or "right" specifying the side to float the element.

Examples

library(htmltools)

div(
  div(
    .style %>%
      border("red") %>%
      float("left"),
    "Warning"
  ),
  div(
    "Nam a sapien.",
    "Phasellus neque orci, porta a, aliquet quis, semper a, massa.",
    "Phasellus lacus."
  )
)

#> <div>
#>   <div class="cas-border-red cas-border-all cas-border-radius-md cas-float-left">Warning</div>
#>   <div>
#>     Nam a sapien.
#>     Phasellus neque orci, porta a, aliquet quis, semper a, massa.
#>     Phasellus lacus.
#>   </div>
#> </div>