Font

The font() function adjusts the size, weight, style, case, and family of the font of a tag element.

font(x, weight = NULL, style = NULL, case = NULL, family = NULL)

Arguments

x
A tag element or .style pronoun.

weight
One of "light", "lighter", "normal", "bolder", or "bold" specifying the font weight, defaults to NULL, in which case the argument is ignored.

If "bolder" or "lighter", the font weight is changed relative to the current font weight.

style
One of "italic" or "normal" specifying the font style, defaults to NULL, in which case the argument is ignored.

case
One of "upper", "lower", or "title" specifying the font case, default to NULL, in which case the argument is ignored.

family
One of "sans-serif" or "monospace" specifying the font family, defaults to NULL, in which case the argument is ignored.

Weights

p(
  .style %&gt;%
    font(weight = "bold"),
  "Curabitur lacinia pulvinar nibh."
)

Curabitur lacinia pulvinar nibh.

p(
  .style %&gt;%
    font(weight = "light"),
  "Proin quam nisl, tincidunt et."
)

Proin quam nisl, tincidunt et.

Examples

library(htmltools)

p(
  .style %>%
    text("indigo") %>%
    font("bold"),
  "Phasellus at dui in ligula mollis ultricies."
)

#> <p class="cas-text-color-indigo cas-font-weight-bold">Phasellus at dui in ligula mollis ultricies.</p>