Understanding responsive arguments

Responsive arguments allow you to apply styles to tag elements based on the size of the viewport (e.g. browser screen). This is important when developing applications for both web and mobile. Specifying a single unnamed value the style will be applied for all viewport sizes. Use the names below to apply a style for viewports of that size and larger. For example, specifying c(default = "center", md = "left") will apply "center" on extra small and small viewports, but for medium, large, and extra large viewports "left" is applied. Styles for larger viewports take precedence.

A responsive argument may be a single value or a named list. Specifying a single unnamed value is equivalent to specifying default or xs. The possible values will be described in the specific help page. Most responsive arguments default to NULL in which case the argument is ignored.

Breakpoints

extra small

Use the breakpoint with default = or sm = .

The style is always applied, unless supplanted by a style for any other breakpoint.

small

Use the breakpoint with small = or sm =.

The style is applied when the viewport is at least 576px wide, think landscape phones.

medium

Use the breakpoint with medium = or md =.

The style is applied when the viewport is at least 768px wide, think tablets.

large

Use the breakpoint with large = or lg =.

The style is applied when the viewport is at least 992px wide, think laptop or smaller desktops.

extra large

Use the breakpoint with \extra-large\verb = or xl =.

The style is applied when the viewport is at least 1200px wide, think large desktops.