Borders

The border() function adjusts a tag element's borders.

border(x, color, sides = TRUE, radius = "medium")

Arguments

x
A tag element or .style pronoun.

color
One of "blue", "indigo", "purple", "red", "orange", "yellow", "green", "teal", "cyan", or "white" specifying the border color.

sides
One or more of "top", "t", "right", "r", "bottom", "b", "left", "l", "all", or "none" specifying which sides to add borders to, defaults to TRUE. TRUE and FALSE may be used as shorthands for all sides or no sides, respectively.

radius
One of "small", "sm", "medium", "md", "large", "lg", "circle", "pill", or "none" specifying the border radius, defaults to "medium".

Colors

blue

indigo

purple

red

orange

yellow

green

teal

cyan

white

Radii

small

sm

medium

md

large

lg

circle

pill

none

Examples

library(htmltools)

h3(
  .style %>%
    border("red", "bottom") %>%
    text("red"),
  "A bold heading!"
)

#> <h3 class="cas-border-red cas-border-bottom cas-border-radius-md cas-text-color-red">A bold heading!</h3>