Floats

The float() function places an element to the left or right side of its parent element. Other text and inline elements wrap around floated elements. Note, float() has no effect on flex items.

Usage

float(x, ...)

Arguments

x

A tag element or .style pronoun.

...

One of the following,

"left"

"l"

"right"

"r"

"none"

Name-value pairs to specify breakpoints.

Value

An object of the same type as x.

Examples

library(htmltools)

div(
  div(
    .style %>%
      border_color("danger") %>%
      float("left"),
    "6.5/10"
  ),
  div(
    "Considering the need for opening sentences.",
    "We may want to reconsider the necessity of second or third sentences.",
    "The whole problem may be avoided by never creating a problem."
  )
)
#> <div>
#>   <div class="border-danger float-left">6.5/10</div>
#>   <div>
#>     Considering the need for opening sentences.
#>     We may want to reconsider the necessity of second or third sentences.
#>     The whole problem may be avoided by never creating a problem.
#>   </div>
#> </div>