Padding

The padding() function adjusts the inner spacing of a tag element. The padding of a tag element is the space between the tag element's border and its content or child elements.

padding(x, all = NULL, top = NULL, right = NULL, bottom = NULL, left = NULL)

Arguments

x
A tag element or .style pronoun.

all
A responsive argument.

One of 1:5 specifying a padding for all sides of the tag element, defaults to NULL, in which case the argument is ignored. 0 removes all inner space and 5 adds the most space.

top, right, bottom, left
A responsive argument.

One of 1:5 specifying a padding for the element's respective side, defaults to NULL, in which case the argument is ignored. 0 removes all inner space and 5 adds the most space.

Panels

Well panels.

wellPanel(
  radioButtons(
    inputId = "id",
    label = "Radio input",
    choices = c(
      "Choice 1",
      "Choice 2"
    )
  )
)

Radio input

Choice 1

Choice 2

Shrink well padding.

wellPanel(
  .style %&gt;%
    padding(1),
  radioButtons(
    inputId = "id",
    label = "Radio input",
    choices = c(
      "Choice 1",
      "Choice 2"
    )
  )
)

Radio input

Choice 1

Choice 2

Auto width.

wellPanel(
  .style %&gt;%
    padding(1),
  radioButtons(
    inputId = "id",
    label = "Radio input",
    choices = c(
      "Choice 1",
      "Choice 2"
    )
  ) %&gt;%
    width("auto")
)

Radio input

Choice 1

Choice 2

Examples

library(htmltools)

div(
  .style %>%
    margin(2) %>%
    border("green") %>%
    padding(2) %>%
    background("red"),
  "Donec vitae dolor."
)

#> <div class="cas-margin-all-2 cas-border-green cas-border-all cas-border-radius-md cas-padding-all-2 cas-background-red">Donec vitae dolor.</div>