Add space into elements

The padding_*() functions adjust the inner spacing of a tag element. The padding of a tag element is the space between the element's border and its content or child elements.

Usage

padding_all(x, ...)

padding_top(x, ...)

padding_right(x, ...)

padding_bottom(x, ...)

padding_left(x, ...)

padding_horizontal(x, ...)

padding_vertical(x, ...)

Arguments

x

A tag element or .style pronoun.

...

One of the following,

0

1

2

3

4

5

Use breakpoints to specify responsive values.

Examples

library(htmltools)

div(
  .style %>%
    margin_all(2) %>%
    padding_all(2) %>%
    border_color("primary") %>%
    background_color("light"),
  "This element has padding"
)
#> <div class="m-2 p-2 border-primary bg-light">This element has padding</div>