Text

The text() function adjusts the text color, alignment, line spacing, line wrapping, and decoration of a tag element.

text(
  x,
  color = NULL,
  align = NULL,
  spacing = NULL,
  decoration = NULL,
  wrap = NULL
)

Arguments

x
A tag element or .style pronoun.

color
One of "blue", "indigo", "purple", "red", "orange", "yellow", "green", "teal", "cyan", "black-50", "white-50", "white", "muted", "body", or "reset" specifying the text color, defaults to NULL, in which case the argument is ignored.

align
One of "left", "right", or "center" specifying the

spacing
One of "small", "sm", "large", or "lg" specifying the text line spacing, defaults to NULL, in which case the argument is ignored.

decoration
One of "none", "underline", or "strikethrough" specifying how the text is decorated, defaults to NULL, in which case the argument is ignored.

wrap
One of TRUE or FALSE specifying if an element's text should wrap onto new lines, defaults to NULL, in which case the argument is ignored.

Colors

div(
  .style %&gt;%
    text("blue") %&gt;%  # &lt;-
    border("blue"),
  p("Nullam tristique diam non turpis.",
    "Pellentesque dapibus suscipit ligula.",
    "Nullam eu ante vel est convallis dignissim."),
  p("Aliquam posuere.")
)

Nullam tristique diam non turpis. Pellentesque dapibus suscipit ligula. Nullam eu ante vel est convallis dignissim.

Aliquam posuere.

Examples

library(htmltools)

div(
  .style %>%
    text(spacing = "small"),
  "Nam vestibulum accumsan nisl.",
  "Fusce commodo."
)

#> <div class="cas-text-spacing-sm">
#>   Nam vestibulum accumsan nisl.
#>   Fusce commodo.
#> </div>

div(
  .style %>%
    text(spacing = "large"),
  "Suspendisse potenti.",
  "Pellentesque tristique imperdiet tortor."
)

#> <div class="cas-text-spacing-lg">
#>   Suspendisse potenti.
#>   Pellentesque tristique imperdiet tortor.
#> </div>

tags$button(
  .style %>%
    text(wrap = FALSE),
  "Aliquam feugiat tellus ut neque."
)

#> <button class="cas-text-nowrap">Aliquam feugiat tellus ut neque.</button>